requirements.txt
torch
transformers
diffusers
hydra-core
torchvision
accelerate
gdown
einops
pytorch-lightning
taming-transformers-rom1504
kornia
